Renamed .js files to .ts. Refactored code in functions/ - added type support
